+++ src/common/zendnn_postops.cpp
@@ -189,7 +189,7 @@ void zenPostOps(
     const int out_width,
     const int no_of_filter,
     const int total_filters,
-    unsigned long biasOffset,
+    unsigned long long biasOffset,
     const float *bias,
     const bool relu,
     const int gelu,
